Move main.go from cmd/server to root directory
The changes appear to be a file relocation and build path updates. The main
server code was moved from cmd/server/main.go to the root directory, with
corresponding updates to build commands in Makefile and air.toml.
