+++ .air.toml
@@ -5,7 +5,7 @@ tmp_dir = "bin"
 [build]
   args_bin = []
   bin = "./bin/server/main"
-  cmd = "go build -o ./bin/server/main ./cmd/server/main.go"
+  cmd = "go build -o ./bin/server/main"
   delay = 1000
   exclude_dir = ["assets", "bin", "vendor", "testdata"]
   exclude_file = []
