+++ Makefile
@@ -1,5 +1,5 @@
 build:
-	go build -o bin/server ./cmd/server/main.go
+	go build -o bin/server
 
 run:
 	go run ./cmd/server/main.go
